So far, Add the "Upcoming Event" pages and the template of article
both Chinese and English, the content have not be changed. Just for
test. The hyperlinks need to be debug. I want to know how to  generate
dynamically so that I can start to write the other pages. There are a
lot of catalogues such as "Review" "Announcement" "Contest Events
&Imformation", and so on. I will give a "read me.txt" to explain my note
in the code later.